Widen size_t / ssize_t / time_t typedefs to long (#52)
c5's bundled headers were carrying `typedef int <name>;` for
every POSIX-byte-count and time-stamp typedef, left over from
before M31 made `int` 4 bytes wide. Result: any libc bridge
that took a `size_t` (`read` / `write` / `mmap` / `lseek`),
`ssize_t` (return values), `off_t` (file offsets) or
`time_t` (clock values) silently truncated to a 4-byte
window. Programs reading large files or computing offsets
past 2 GiB / timestamps past 2038 saw the wrong number on
its way back; the c5 stack slot is 8 bytes either way, but
the type's _stored_ width is what the codegen uses for
sign-extending loads (`Lwu` vs `Lw`) and what `sizeof`
reports.

Three header touch-ups:

  * `<stddef.h>` -- `size_t` and `ptrdiff_t` are now `long`
    (8 bytes on every host this dialect targets). `wchar_t`
    stays at `int` because UTF-32 codepoints fit there and
    no host expects a wider value.
  * `<sys/types.h>` -- `ssize_t`, `off_t`, `off64_t`,
    `loff_t`, `ino_t`, `ino64_t`, `blkcnt_t`,
    `blkcnt64_t`, `fsblkcnt_t`, `fsfilcnt_t` -- all
    promoted to `long`. The 4-byte-everywhere identity
    types (`pid_t`, `uid_t`, `mode_t`, ...) keep `int`.
  * `<time.h>` -- `time_t` and `clock_t` -> `long`. The
    Y2038 horizon is closer than the typedef tweak's
    follow-up.

Promoted the previously-deferred fixture
`fixtures/c/deferred_width_typedefs.c` and its companion
`width_typedefs_truncated_to_int` test. The test now runs
unconditionally and passes (asserts `sizeof(size_t) ==
sizeof(void*)` etc.).

Cargo: 349 tests pass + 10 ignored, fmt clean.
sqlite3 in-memory smoke still gives `-7|neg / 1|hello /
2|world / 3|-4` on macos-aarch64.

Per-target `long` width (Windows LLP64 vs LP64) stays open
as #51; today every c5 target uses `long = 8` so the
typedefs already match macOS/Linux. Windows would need
`long long` to land alongside `long` for the same
pointer-wide promise; deferred separately.

+++ headers/include/stddef.h
@@ -1,14 +1,20 @@
 // stddef.h -- the standard "common definitions" header.
-// c5 collapses every integer to a 64-bit signed word, so size_t /
-// ptrdiff_t / wchar_t all alias `int`. NULL is the canonical zero
-// pointer literal. The offsetof shape `&((T*)0)->m` is recognised
-// by the constant-expression evaluator (see parse_const_offsetof
-// in compiler.rs); the macro form below is the customary expansion.
+// `size_t` and `ptrdiff_t` are pointer-wide on every 64-bit
+// host (LP64 macOS / Linux: 8 bytes; LLP64 Windows: 8 bytes
+// pointer but 4-byte `long`; we still alias `long` and accept
+// the small Windows skew per #51). `wchar_t` stays at `int`
+// width since UTF-16 / UTF-32 codepoints fit in 4 bytes and
+// no host expects a wider value here.
+//
+// NULL is the canonical zero pointer literal. The offsetof
+// shape `&((T*)0)->m` is recognised by the constant-expression
+// evaluator (see parse_const_offsetof in the compiler); the
+// macro form below is the customary expansion.
 #ifndef _C5_STDDEF_H
 #define _C5_STDDEF_H
 
-typedef int size_t;
-typedef int ptrdiff_t;
+typedef long size_t;
+typedef long ptrdiff_t;
 typedef int wchar_t;
 
 #ifndef NULL
